+++ cardano-crypto-tests/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## 2.2.0.1
 
-*
+* Add tests using standard test vectors and generated ones for Praos and PraosBatchCompat
 
 ## 2.2.0.0
 
@@ -38,4 +38,3 @@
 ## 2.0.0.1
 
 * Initial release
-